AI-Image-Style-Transfer-Tool/requirements.txt
transformers
torch
flask
gradio
numpy
pillow
diffusers
accelerate